modulo de plantillas
