+++ plantillas/controladores/seleccionarPlantillaControlador.js
@@ -1,44 +1,36 @@
-// Se importa un módulo personalizado que establece la conexión con la base de datos
-const conexion = require('../../util/bd.js');
+const { obtenerPlantillaRepositorio } = require('../data/repositorios/repositorioSeleccionarPlantilla.js');
 
-// Se exporta una función asíncrona que maneja la consulta de plantillas
-exports.seleccionarPlantilla = async (pet, res) => {
-    // Se extrae el idPlantilla del cuerpo de la petición
-    const { idPlantilla } = pet.body;
+/**
+ * Controlador para seleccionar una plantilla por su ID
+ */
+exports.seleccionarPlantilla = async (req, res) => {
+  // Se extrae el idPlantilla del cuerpo de la petición
+  const { idPlantilla } = req.body;
 
-    // Si no se proporciona idPlantilla, se devuelve un error 400 (Bad Request)
-    if (!idPlantilla) {
-        return res.status(400).json({
-            message: 'Faltan datos requeridos',
-        });
+  // Si no se proporciona idPlantilla, se devuelve un error 400 (Bad Request)
+  if (!idPlantilla) {
+    return res.status(400).json({
+      mensaje: 'Faltan datos requeridos',
+    });
+  }
+
+  try {
+    // Se obtiene la plantilla desde el repositorio
+    const plantillaConsultada = await obtenerPlantillaRepositorio(idPlantilla);
+    
+    if (plantillaConsultada.length === 0) {
+      return res.status(404).json({
+        mensaje: 'Plantilla no encontrada',
+      });
     }
 
-    // Se llama a una función auxiliar para obtener la plantilla desde la base de datos
-    const plantillaConsultada = await obtenerPlantilla(idPlantilla);
-
-    // Se responde con un mensaje de éxito y se devuelve la primera plantilla encontrada
+    // Se responde con un mensaje de éxito y se devuelve la plantilla encontrada
     res.status(200).json({
-        message: 'Consulta de plantilla exitosa',
-        plantilla: plantillaConsultada[0], // Se asume que solo se necesita el primer resultado
+      mensaje: 'Consulta de plantilla exitosa',
+      plantilla: plantillaConsultada[0], // Se devuelve el primer resultado
     });
-}
-
-// Función auxiliar que realiza la consulta a la base de datos
-async function obtenerPlantilla(idPlantilla) {
-    // Se retorna una promesa que ejecuta una consulta SQL usando el idPlantilla
-    return new Promise((resolver, rechazar) => {
-        const consulta = 'SELECT * FROM plantillareporte WHERE idPlantillaReporte = ?';
-
-        // Se ejecuta la consulta con el valor de idPlantilla como parámetro para evitar inyecciones SQL
-        conexion.query(consulta, [idPlantilla], (err, resultados) => {
-            if (err) {
-                // En caso de error, se muestra en consola y se rechaza la promesa
-                console.error('Error al ejecutar la consulta:', err);
-                return rechazar(err);
-            }
-
-            // Si no hay error, se resuelve la promesa con los resultados obtenidos
-            resolver(resultados);
-        });
-    });
-}
+  } catch (error) {
+    console.error('Error al seleccionar plantilla:', error);
+    res.status(500).json({ mensaje: 'Error interno del servidor' });
+  }
+};
