+++ plantillas/rutas/eliminarPlantillas.ruta.js
@@ -1,5 +1,3 @@
-//RF21 Usuario elimina plantilla de reporte. - https://codeandco-wiki.netlify.app/docs/proyectos/tractores/documentacion/requisitos/RF21
-
 const express = require('express');
 const ruteador = express.Router();
