+++ plantillas/rutas/seleccionarPlantilla.ruta.js
@@ -1,5 +1,3 @@
-//RF23 Usuario consulta plantillas de reporte. - https://codeandco-wiki.netlify.app/docs/proyectos/tractores/documentacion/requisitos/RF23
-
 const express = require('express');
 const ruteador = express.Router();
